youtube_summary_gitaction/requirements.txt
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
youtube-transcript-api>=0.6.0
google-generativeai>=0.8.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
